Update patch set 4
Patch Set 4:

PS4 reverts a bad test in the compiler driver when deciding if the profile should be use. The presence of the profile file should be enough to decide if the profile data should be used. Using runtime options adds unnecessary complication when dex2oat is started from class linker.

Patch-set: 4
Label: Verified=0