Update patch set 1
Patch Set 1:

> I'm OK to only deal with jobject. We just have to decode them to access the Object* (which we already do). For the map key, I think we should evaluate the cost and compare to what we have now.

Right.

> There is a mechanism where the GC notifies the debugger about system weaks so we wait before adding any new system weak in the object registry: see ObjectRegistry::Disallow/AllowNewObjects. But this does not prevent from accessing the existing weaks in the object registry.

Yes. I think if we use jobject as the map key, in addition to disallowing new system weaks for that time period, we'd need to disallow access to the existing weaks at the same time. Is this a deal breaker?

I'm currently leaning toward trying out the identity-hash-code-as-the-map-key approach.

Patch-set: 1
Label: Verified=0