Update patch set 3
Patch Set 3:

(1 comment)

I wonder what's the cost of calling ArtMethod::GetInterfaceMethodIfProxy each time we access the DexFile::CodeItem or the method name. We did not have to do that each time as it was stored in MethodHelper::method_ on (re)initialization.

The good point is we may no longer need to pass the MethodHelper and DexFile::CodeItem as parameters to the interpreter because the ArtMethod would give us direct access to such information. This could be done later.

Patch-set: 3
Label: Verified=0