Update patch set 2
Patch Set 2:

No, I don't have a design document. I was just planning to write tests to show what I expect the behavior to be.

Eliminating unnecessary operations in ME would be the next logical step. I'll do that in a subsequent CL if time allows. I didn't think about the BE in this context.

Patch-set: 2
Label: Verified=0