Update patch set 1
Patch Set 1:

(1 comment)

I guess the question was for both: whether they should be, and how.

But it currently is how irogers wants it: -libart is tied into the host test, but not the run-test.

The question will become more interesting once we have multilib tests on the host. The issue with the target is that the non-qualified Make rule is forwarded to the primary-arch target rule, so there isn't anything to "link in" the libart rule.

Patch-set: 1