Update patch set 17
Patch Set 17:

Hiroshi,

This change includes entrypoints for all the native StringFactory allocation methods. JNI overhead is definitely an issue with this change, so thanks for pointing that out. There are still issues with runtime performance that I am working on. Ritzperf is actually faster with this change (~3200 vs ~3600 previously), but String in Caffeinemark is worse.

Patch-set: 17
Label: Verified=0