Update patch set 9
Patch Set 9:

Although the solution in this patch is great, the implementation does not match the pre-existing methodology for check elimination. Consider MIRGraph::BasicBlockOpt. Check elimination is done by marking a MIR's optimization_flags with MIR_IGNORE_SUSPEND_CHECK. By doing it this way, it allows the middle-end to be aware of the check elimination and it minimizes changes to backend because it already has logic to handle the check elision via that flag. The target independent part of check elision should have been implemented as a compiler pass that marks MIRs with the appropriate flag. This would have prevented the changes in mir_to_lir.cc. Ian, Vladimir, Wei, what do you think about this approach and will you reconsider it before turning this on by default?

Patch-set: 9
Reviewer: Razvan A Lupusoru <1013855@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0