Update patch set 1
Patch Set 1: Code-Review-1

Currently we profile using suspend checks which are added before the method return. So capturing the program counter is not particular useful IMO as it will always be the same. (i.e. you can't detect a function hot spots with this).

What is the use case that you have in mind or what trigger the need to have the PC?

Patch-set: 1