Update patch set 2
Patch Set 2:

PS2 shows the approach that Ian and I do favor. We want to use the generic JNI and a new modifier flag instead of an additional pointer per method.

With a little abuse of the current API, this actually runs a "native bridge" (which itself just runs native code) in run-test 112.

I have a couple of comments for the API that I will forward tomorrow.

Patch-set: 2