Update patch set 1
Patch Set 1:

Nicolas,

Would you mind taking a look at the live_ranges_test.  This is failing because the 'equal' node is no longer present for the 'if' instruction.  This CL removes that node and puts a condition in its place, thus generating better code for all the IF_XX instructions.  The live_ranges_test assumes some numbers I am not familiar with and don't know what the output should be.

I have modified the liveness_test to remove the 'equal' instruction from the bitsets so it's passing now.

Patch-set: 1