Update patch set 3
Patch Set 3:

Bill, here is my thoughts. current vreg map encoding table allows you to separate fp, core regs and frame. So in reality we need to distinguish 32 bit, 32-bit high and 32-bit low. User of vreg table encoding table knows what type of register it is interested in. So it can distinguish them by providing hint (as actually it does right now).
The only problem is actually when on the same architecture long or double can be promoted to both solo 64-bit reg and pair of 64-bit regs. Do we really need this possibility? In this case we really need some additional information in vreg table. I thought about this approach but it will increase the size of vreg table while I'm not really sure that we need this case.

Patch-set: 3
Label: Verified=0