Update patch set 2
Patch Set 2:

(5 comments)

Thanks for adding this. It was on our todo list, together with the extended indexed load/store. Notice also the comments in

  https://android-review.googlesource.com/#/c/98638/

I suspect we could move the functions EncodeExtend(), EncodeShift(), IsEncodedExtend() to codegen_arm64.h for inlining purposes (not necessarily in this commit).

Patch-set: 2
Label: Verified=0