Update patch set 1
Patch Set 1: Code-Review+2

And a (weak) suggestion, feel free to keep as, but when I first encountered "is_volatile" I expected it to be boolean.  Another possibility would be to rename it "volatility", which in my usage would be more in keeping with having states of volatile and not volatile.  But again, I'm also fine with it staying the way it is.

Patch-set: 1
Reviewer: Bill Buzbee <1001578@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2