Update patch set 2
Patch Set 2:

The context is that some test configurations are causing machines to lock up. In this circumstance optimizing run-tests with multilib enabled. The point of the ulimit is to avoid errant behavior. If a test doesn't fit within the ulimit then it would be broken today. For new tests 2MB is pretty generous, but we can also but the limit on a test by test basis. The point is not to run the test until all the resources on your machine are consumed, the file limit being just part of that.

Patch-set: 2