Update patch set 1
Patch Set 1:

What fails with the interpreter?  I did upload a change to switch the implicit runtime checks off when running in interpreter only mode (that is, don't register the signal handlers).  However if you are seeing the interpreter fail when implicit checks are on it might be a problem since we do enter the interpreter at other times (deopt, debug, etc.)

Patch-set: 1
Label: Code-Review=0