Update patch set 1
Patch Set 1:

Nicolas, I met this issue accidentally and I do not know how to write a test to reach a buggy condition. Specifically, you need that register allocator comes to array out of bound exception code with the index in specific registers. So I doubt I can provide a test case for this bug.

Patch-set: 1