Update patch set 1
Patch Set 1: -Code-Review

> Nice catch Elena! By any chance, were you able to reproduce the bug
 > in a self-contained test?

I have smaller test created from the reliability one.
For me the issue is reproduced on it after N min.
So the test is more “stress” than unit.
I may put it in the comment if ~80 lines of code is appropriate here. 
Or you are fine with the reliability one?

Patch-set: 1
Label: Code-Review=0