Update patch set 2
Patch Set 2:

As a follow-up to this CL, it's discussion-worthy what to do with TargetReg. The extension was necessary to get specific views of special registers, and because TargetReg leaves it open at the moment what view will be returned.

It seems all backends are gravitating towards 32b views from TargetReg, which, if we specified, we could use instead of TargetReg(..., false). Or we could drop the unspecific TargetReg completely.

Patch-set: 2