Update patch set 2
Patch Set 2:

Personally, I am in favour of removing the ambiguity in the 1-argument version of TargetReg(). I have no strong opinions about how we achieve this (whether we do this by removing the 1-argument version of TargetReg() altogether or by making it return always a 32-bit view). A possibility would be introducing TargetReg32(regKind) and TargetReg64(regKind), just to avoid having these (true, false) arguments floating around in the code. I would still keep TargetReg(regKind, isWide), as it is often nice to be able to parametrize the code with respect to the wideness (e.g. keep it for all the cases where isWide is not a known boolean constant).

Patch-set: 2