Update patch set 1
Patch Set 1:

(20 comments)

Thanks for the major cleanup, you did my todo for tomorrow. :-)

General note: I am not obsessed to bring the TargetReg into the backends. I assume you know what you are doing there and don't need to go through the hoop of a SpecialTargetRegister. So if you find out that the compiler performance is impacted negatively, or that for you readability is better the old way, I'm fine with using the explicit reg storage definitions...

Will also add ARM people to the reviewer list as they are impacted by the general changes, and there are one or two comments that are general questions.

Patch-set: 1
Label: Code-Review=0