Update patch set 2
Patch Set 2: Code-Review+1

I’ve added a test which reproduced the issue during 3 times out of 3 for me in ~1min. Note that thread suspend timeout only happens after 30 sec. Please verify if it may be added to the art test or please use in other test suite. This is the smallest test which somewhat stably reproduces the issue for me.

Patch-set: 2
Label: Code-Review=+1