Update patch set 1
Patch Set 1: Code-Review-1

I believe the EvalLoc() and LoadValue() are working as intended. Both force the locations into a register, grabbing an _empty_ temp if necessary. The overlap is _usually_ (more below) possible only if both rl_src and rl_dest are the same _promoted_ vreg. In that case, the rl_src will contain the old value after the EvalLoc()/LoadValue() sequence and that can be edited in place, creating a new value that will be in the vreg and the final StoreValue() will be a NOP.

The only problem I see is the strange situation when we're generating an intrinsic and the target is an invalid location (i.e. not a vreg). If the rl_src is in the default return location we shall wrongly overwrite the source. We've dealt with this before in https://android-review.googlesource.com/98492 and we should deal with it the same way for the rest of the intrinsics. Maybe that fix should go directly to the inliner to avoid putting the burden on architecture-specific codegens.

Patch-set: 1
Reviewer: Vladimir Marko <1018108@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1