Update patch set 2
Patch Set 2:

Bill,
Although this CL may seem significant, it actually is not. It mostly moves frontend code to its own class. The significant method it adds is MergeCFG which basically does what the MIRGraph::InlineMethod was supposed to do. This patch passed our internal QA patch testing and thus I expect the merge risk is low. I will rebase on a more recent AOSP head to make it ready for a merge.

Patch-set: 2
Label: Verified=0