Update patch set 8
Patch Set 8:

> So will we patch the boot file and the application file at startup
 > of the VM? What happened to avoiding dirty pages? Or will we create
 > another file?

patchoat creates another file in the usual place in /data.

so this is an option for devices that want to burn more space (since they need the file in /system and /data) to get faster out of the box (and factory testing).

it also means we can spend more time compiling the contents of /system since it is done offline.

it also means when /data in tmpfs (booting minimal UI to get /data decryption key) we don't have to dex2oat but just relocate.

Patch-set: 8