Update patch set 1
Patch Set 1:

I'm not sure it achieves it. The current patten is to just copy an entire directory and then insert code as appropriate. The new code confuses this by sharing files, not having golden files and introducing choice. I did a round of trying to move oat tests into run-tests as run-tests have GDB support and other features. I'd like there to be fewer not more ways of doing tests in ART, but I'm happy to see the testing cleaned up. Currently if you want an execution result true/false test then it should be an oat test. This is important when there non-deterministic output, but that's not the case with the test case you are adding.

Patch-set: 1