Update patch set 1
Patch Set 1:

> Aside from non-deterministic output and using a exit code to confer
 > success, the oat tests do an explicit dex2oat rather than relying
 > on the runtime to fork it. This can be useful for debugging dex2oat
 > and so we should probably make it a run-test option. Note that the
 > dex2oat in the oat tests is brittle and if the generated file has
 > the wrong name the run-time will silently compile the file again.
 > In the worst instance this can make you think you are testing an
 > optimizing compiled thing whereas the code was really compiled with
 > quick.

oat tests also seem to use the build system's logic for compiling Java files, and the use of private API in these tests make it so that the Java compilation needs to be special cased also if we want to make them run-test.

So how do you want to move forward with this change? Existing run-test can't really be changed because of their golden test nature. In a future CL, we can move oat tests to run tests, and this change is just making it easier to make this move.

Patch-set: 1