Update patch set 1
Patch Set 1: Code-Review-1

Yes, the original (coming from a recent Mips change according to git blame) is better with respect to ref, as ref returns on 64b should be 64b registers.

I was surprised when seeing that we seemingly never hit this slow-path on x86 or ARM64.

Patch-set: 1
Label: Code-Review=-1