Update patch set 1
Patch Set 1:

I think that would add complexity to the code base that is probably not worth it. That would weaken the invariants/checks we do around the code, which makes the maintenance/debugging harder. As a side story, during the development of Jeff's new string work, the rosalloc verification detected a pre-fence barrier bug as it resulted in a racy case where some objects' size looked too short for a slot, which would not be detected without those invariants. I'd rather keep things simple. With the OOM-time compaction available and the suggested approach would not achieve true 100% compaction (ie. as tight as the bump pointer space), the upside does not seem a lot to me, at least for now.

Patch-set: 1