Update patch set 1
Patch Set 1:

> You want lots of tests because you want compiler regressions to be
 > the norm?

Not sure I follow. It's not about compiler regressions here, it's about tests for the VM, regressions or not. The more, the merrier. And the easier it is for VM engineers and external contributors, the better.

I've seen some CLs where testing some functionality was added to existing tests. I think this is a problem, and our current setup is not making it simple to just create a new test.

Patch-set: 1