Update patch set 1
Patch Set 1:

If there is a test, whether it appears as part of 1 file or as a stand-alone test, its still just the contribution of 1 test. Adding files doesn't bring about extra test coverage. My concern in making every bug a test is that we're going to explode the number of test cases to the point that folks won't run as many tests.
Could we make this class of bug a test case within run-tests and then make the infrastructure local to that bug? That way new tests can be new files but we avoid an explosion in the number of test cases.

Patch-set: 1