Update patch set 1
Patch Set 1:

> > If there is a test, whether it appears as part of 1 file or as a
 > > stand-alone test, its still just the contribution of 1 test.
 > Adding
 > > files doesn't bring about extra test coverage.
 > 
 > When I mentioned coverage, I talked about running multiple tests in
 > the same test. So if you have one failure in that test, you won't
 > run the remaining ones.
 
But we shouldn't have failures, so why optimize for that case? Its a burden in the mechanics to do so.

 > My concern in making
 > > every bug a test is that we're going to explode the number of
 > test
 > > cases to the point that folks won't run as many tests.
 > 
 > Yeah, I share that concern as well.
 > 
 > So there's two things in play here:
 > - We want lots of tests
 > - We don't want to run all these tests before each commit
 > 
 > And two solutions:
 > - Make running tests faster (with faster VM startup, faster
 > compilation, faster infrastructure, etc).
 > - Have a sanity test "suite" that everyone should run, and then
 > rely on the buildbot to tell us if we're green on the rest.
 
No. We should run as many tests pre-commit as possible. The buildbot is just an honesty test.

 > > Could we make this class of bug a test case within run-tests and
 > > then make the infrastructure local to that bug? That way new
 > tests
 > > can be new files but we avoid an explosion in the number of test
 > > cases.
 > 
 > Are you suggesting to not make it part of, eg, test-art-host? That
 > works for me, as long as the tests are being run by the buildbot.

No. I'm suggesting that as run-test just calls run in a test, you can get the behavior you are after by pushing the infrastructure into a test. So don't modify run-test, create a run script that does what you want. At some later date we can go around and sanitize the tests, remove duplicates, etc..

Patch-set: 1