Update patch set 1
Patch Set 1:

Sorry, I kept the description short. Here's the long one.

When we're dying, we try to dump the Java stack. In the process, we also try to dump held locks, e.g, when it's a synchronized method. We do this by trying to print the identity hash code of the monitor object, as this is more meaningful than the pointer itself.

Now consider the case where the monitor object is only thin-locked. Then computing the identity hashcode will inflate the monitor. The monitor would like to know the method and pc creating it, as we normally might want to log that for contention analysis.

Finding the current method is another stack visitor. This visitor is the one that is modified - instead of failing, it will now delegate handling bad PCs to the caller. This CL elects to ignore bad locking PCs as occurring when a fault happened. The only downside is that we could spuriously have invalid line numbers when doing contention logging (though I would expect that we'd die somewhere else anyways).

Patch-set: 1
Label: Verified=0