Update patch set 6
Patch Set 6: Code-Review+2

(3 comments)

Does this pass the assembler_thumb_test?  That is still not enabled in the main build and needs to be manually added to the makefile due to issues finding the objdump command on the build servers.  I can take a look at that.

Patch-set: 6
Reviewer: Dave Allison <1036486@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2