Update patch set 1
Patch Set 1:

I am not sure we can even do a best effort approach.  The problem is that when the SEGV occurs and this is called we already know that it is not one of our signals so this is a crash.  With a crash we cannot know the state of anything.  The whole stack may be completely invalid.  Since this is not a debugger (using ptrace) if we try to dereference an invalid address we will get another segv.  

Do we not get a stack trace anyway from the signal catcher?

Patch-set: 1
Reviewer: Dave Allison <1036486@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0