Update patch set 2
Patch Set 2:

Fixes b/12437401.

This is lacking a test, but I still wanted to send it to show how simple a change it is, and make sure that is what we want. For example, do we want a specific crash log to detect those failures? Send a message to the developers (can we do that with just an apk?), etc.

Patch-set: 2
Label: Verified=0