Update patch set 2
Patch Set 2: Code-Review+2

This is better than what we have but we shouldn't be adding multiple notions of stack end, especially when it makes work for the interpreter. We should clean up the tls values to hold this value if necessary, but given we're rewriting this code at the moment let's look to do that as part of the rewrite.

Patch-set: 2
Reviewer: Ian Rogers <1010118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2