Update patch set 2
Patch Set 2:

oat tests are dying, but until then run them properly.

This typo explains why test-art-host-oat-optimizing-SignalTest32 whould work on a 32bit host build and fail on a 64bit host build: because dex2oat does not generate the code in the expected path (x86_64 instead of x86), dalvikvm would re-generate it, with the defualt settings (that is with quick compiler), and hence the test would not fail.

Patch-set: 2
Label: Verified=0