Update patch set 1
Patch Set 1:

The documentation is describing the Dalvik implementation which was wrong. The intent of freeMemory is to give you an idea of how much memory you can allocate in the future. For something like:

int free = (int)Runtime.getRuntime().freeMemory();
byte[] bytes = new byte[free - 4096];  // leave a page of slop at the end

Total memory is a high water mark of the memory used in the past and doesn't relate to the size of allocation that would succeed. As such the Dalvik implementation was less useful for application developers. Its also painful to compute as it requires us to analyze the size of many parts of the heap generally under locks.
I believe the inconsistency you are reporting is a different issue where allocated + free is greater than total. That can have its own test case.

Patch-set: 1