Update patch set 1
Patch Set 1:

Oh I see. Thanks for the clarification. Here are my two cents in case it's useful:

I think that the intended meaning of the three Runtime fields is quite clear: maxMemory is how much it can allocate (which if not set it can even be Long.MAX_VALUE), totalMemory is how much is currently allocated by the VM and freeMemory is how much of the totalMemory is free. 

Yes, the spec is very loose in that regard but it says:

totalMemory = amount of memory for current and future objects
freeMemory = an approximation to the amount of memory for future objects

So it is clear that totalMemory includes freeMemory, and in fact earlier versions of the spec explicitly said that totalMemory was >= than freeMemory : https://www.cs.princeton.edu/courses/archive/fall97/cs461/jdkdocs/api/java.lang.Runtime.html#freeMemory()

But that part was removed because it was difficult to ensure mainly due to race conditions, not because freeMemory is not related to totalMemory: https://bugs.openjdk.java.net/browse/JDK-4057982

Patch-set: 1
Reviewer: Esteban de la Canal <1047453@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0