Update patch set 4
Patch Set 4:

The test is currently running only on the host. The setup is complicated: libcore code needs to find libarttest, but host dlopen must fail to load it. This is currently done by copying the libraries into the test directory and using that directory for the java.library.path. On the host, we can set all of that up "easily" in a custom run script, but for the target that is more complicated.

But I think even in this state this is enough to demonstrate that at least the basic features of the interface work.

Patch-set: 4
Label: Verified=0