Update patch set 5
Patch Set 5:

So the reason for implementing the calculator is because when the x86_64 support was introduced we found that the length of some instructions were incorrectly calculated.  Rather than chasing our tails and finding the instructions one at a time, I thought we should do it properly with a function that will work for any instructions we might encounter in the future.

I don't understand why you would prefer to hack something together rather than using the canonical information from the manual to get the right answer every time.

The old code was basically doing the same thing as this, only incorrectly.  Fixing it was ad-hoc and error prone.

If you don't like the python script we can certainly remove it and use its output as the data for the calculator.

Patch-set: 5