Update patch set 1
Patch Set 1:

(4 comments)

I don't think the unspill can be optimized. The frame size is always larger than the spill size, as we spill the ArtMethod reference to the stack. So you need to start with a pre-indexed load. But then all other loads will also need to be pre-indexed, which means you do not end up at the top in the end. Some combined displacement + post-index would be nice. Can you make the change to make this more convenient? ;-)

Patch-set: 1
Label: Verified=0