Update patch set 1
Patch Set 1:

(1 comment)

You are right about optimising the unspill sequences. I had a better look at the problem and understood that it is not as easy as I thought. I guess that with the current frame layout the way we currently do the unspills is probably the optimal one.

Patch-set: 1
Label: Code-Review=0