Update patch set 3
Patch Set 3:

I guess the increase in size of boot.oat may be due to avoiding promoting constants that would be normally promoted without this patch. After all a single "ldr" is half the size of a couple of "movz/movk", but this doesn't mean that the "ldr" is preferable. Can I ask how much this is?

Patch-set: 3