Update patch set 3
Patch Set 3:

This is in line with what I saw. This reduces typically one store in the entry sequence and one in the exit sequence, but trades that for a store to the stack. That store is necessary, as it was setting a dalvik register - which needs to be visible in case of exceptions etc.

I was talking to Bill before and was suggesting to change the counting for constants and non-constants. My example was to multiply all current values by 10, and make constants count as one. Then a constant in a loop could still be promoted, which is beneficial if we have enough registers (e.g., on 64b architectures). But we were not sure whether we should make such a change this late in development.

Anyways, I'll likely take the patch, the previous reply was just an update as I did not get around to more testing.

Patch-set: 3
Label: Verified=0