Update patch set 1
Patch Set 1:

(7 comments)

Testing of the functionality you do in JNI_Onload is good. I don't really see the point for testing trampolines, though, as the native bridge interface doesn't require trampolines (having it in the name in the API is maybe a bit unlucky) - from ART's side it doesn't matter whether it's trampolines or direct pointers.

But it's OK. Please fix the nits and I'll take it.

Patch-set: 1
Label: Verified=0