Update patch set 1
Patch Set 1:

And for trampoline test, though we don't really need a trampoline here, as they are the same ISA. But conceptually this is how native-brdige (should) work. And with the trampoline we can do something extra inside downcall, like what we are doing in the JNI_OnLoad, without breaking the original JNI functions.

Patch-set: 1