Update patch set 3
Patch Set 3:

OK, thanks for the clarification. I guess the only way to avoid the spurious stores is to extend the runtime so that it knows which vregs contain constants at each suspend point and which are the values of these constants. But it is clearly too late for this.
Experimenting with promotion weights is definitely a good thing to try. I plan to do it as well now that load/store elimination is enabled. I have indeed the suspicion that register promotion is often keeping load/store elimination from working effectively.

Patch-set: 3