Update patch set 1
Patch Set 1:

(2 comments)

I thought that different kinds of checks in the loop is the main overhead in the java version, so I didn't consider too much on writing different implementation for arm and arm64.

Thanks for your comment, especially on the NullCheckElimination which I am not familiar with. I will rewrite them in arm and arm64 child classes.

Patch-set: 1
Label: Verified=0