Update patch set 2
Patch Set 2:

I talked to Nicolas about this tonight and he was comfortable with it. 

We actually have some attempts to be clever in the OatHeaderSizeCheck test. I think we need something similar for the image version number (which actually is what needed to change here) so that if people change the mirror class layout, it will remind them to bump the version number.

Patch-set: 2