Update patch set 7
Patch Set 7:

> (6 comments)

Thanks for working on this! It's great to see we can set breakpoint in Java code.

But like Alex, I am a bit concerned about the approach to set address here.

I'll try to explain how symbol address is formed from the beginning. Alex: please correct me if I'm wrong :-)

There are 3 stages:

1. Compiler. This happens in compiler/oat_writer.cc. Here symbol address is set to quick_code_start, which is basicly offset from this method to the beginning of .text section in the .oat file.

2. Patchoat. Compiler does not generate position independent code; which means we need some data/code to be at certain position. Patchoat is written to address this issue and only this issue. In this stage, symbol address is still offset to beginning of .text section

3. Runtime. This happends in runtime/elf_file.cc. Here we add symbol address by text_start, which is virtual address of .text section. So now symbol address is the actual code address in memory.

So yes, we already know the correct symbol address in stage #2, because text_start is already known after .oat file is generated; but let's keep patchoat doing one thing only, and not complicate that.

In #1 & #3, I still prefer to keep it in #3 for now. If you really want to make it happen in #1, you should discuss with some folks first and then submit the patch.

Patch-set: 7
Label: Verified=0