Update patch set 7
Patch Set 7:

> Alex:
 > I'm not fully agree with your decision. And before writing a bunch
 > of new code to make this patch accepted, I'd like you to confirm
 > that you understand that:
 > 1. I removed all the .debug_abbrev and .debug_line parsing code
 > (about 260 lines in elf_file.cc) and added .rel.debug (30 lines in
 > elf_file.cc and about 15 lines in elf_writer_quick.cc).
 > 2. Without .rel.debug I will need to add another one parser for
 > .debug_line (Line Table Program, DWARF spec Section 6.2).
 > 3. All this parsing is just for fixing address offsets in the debug
 > sections.

Thanks for getting back to me about this. 

I am going to spend some time trying to make this be done my way, and if I'm not able to get something better by the time this is ready to merge I will not stop it with the current design. 

I still do think that going to this weird .rel.debug is not wise though.

Patch-set: 7